+++ compiler-core/src/parse.rs
@@ -699,17 +699,6 @@ where
                     }
                     _ => {}
                 }
-                if tail.is_some()
-                    && elements.is_empty()
-                    && elements_after_tail
-                        .as_ref()
-                        .is_none_or(|elements| elements.is_empty())
-                {
-                    return parse_error(
-                        ParseErrorType::ListSpreadWithoutElements,
-                        SrcSpan { start, end },
-                    );
-                }
 
                 match elements_after_tail {
                     Some(elements) if !elements.is_empty() => {
@@ -3565,17 +3554,6 @@ where
                         );
                     }
                     _ => {}
-                }
-                if tail.is_some()
-                    && elements.is_empty()
-                    && elements_after_tail
-                        .as_ref()
-                        .is_none_or(|elements| elements.is_empty())
-                {
-                    return parse_error(
-                        ParseErrorType::ListSpreadWithoutElements,
-                        SrcSpan { start, end },
-                    );
                 }
 
                 match elements_after_tail {